requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
openai>=1.30.0
azure-cognitiveservices-speech>=1.35.0
pydantic==2.6.0
requests==2.31.0
psycopg2-binary>=2.9.6
authlib==1.4.0
httpx==0.28.1
PyJWT==2.8.0